Fix minor compilation warnings in C/C++ interface. Fix backwards compat in use of unicode Python C API for Python 2. Change assert_result_equal method on test class for MetaschemaTypes to class method to allow call from test_types in Python 2. Change how library path for the Python C API is determined on Windows.

+++ yggdrasil/languages/C/datatypes/AsciiTableMetaschemaType.h
@@ -330,7 +330,7 @@ public:
     if (ret < 0) {
       ygglog_error("AsciiTableMetaschemaType::deserialize: Error using table.");
       return -1;
-    } else if (ret != nargs_exp()) {
+    } else if ((size_t)ret != nargs_exp()) {
       ygglog_error("AsciiTableMetaschemaType::deserialize: Table used %d arguments, but was expected to used %d.",
 		   ret, nargs_exp());
       return -1;
